refactor: reorganize module exports, create `auth` and `types` submodules
